Preparation for moving WaitForNodes into deploy
part of #390
