+++ pkg/util/managedcluster/managedcluster.go
@@ -1,15 +1,19 @@
 package managedcluster
 
 import (
+	"context"
 	"io/ioutil"
 
 	"github.com/ghodss/yaml"
 	"k8s.io/client-go/kubernetes"
+	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/clientcmd"
 	kapi "k8s.io/client-go/tools/clientcmd/api"
 	"k8s.io/client-go/tools/clientcmd/api/latest"
+	"k8s.io/client-go/tools/clientcmd/api/v1"
 
 	"github.com/openshift/openshift-azure/pkg/api"
+	"github.com/openshift/openshift-azure/pkg/util/wait"
 )
 
 func ReadConfig(path string) (*api.OpenShiftManagedCluster, error) {
@@ -41,4 +45,48 @@ func ClientsetFromConfig(cs *api.OpenShiftManagedCluster) (*kubernetes.Clientset
 	}
 
 	return kubernetes.NewForConfig(restconfig)
+}
+
+// getKubeconfigFromV1Config takes a v1 config and returns a kubeconfig
+func getKubeconfigFromV1Config(kc *v1.Config) (clientcmd.ClientConfig, error) {
+	var c kapi.Config
+	err := latest.Scheme.Convert(kc, &c, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	kubeconfig := clientcmd.NewDefaultClientConfig(c, &clientcmd.ConfigOverrides{})
+
+	return kubeconfig, nil
+}
+
+// ClientSetFromV1Config takes a v1 config and returns a Clientset
+func ClientSetFromV1Config(ctx context.Context, config *v1.Config) (*kubernetes.Clientset, error) {
+	kubeconfig, err := getKubeconfigFromV1Config(config)
+	if err != nil {
+		return nil, err
+	}
+
+	restconfig, err := kubeconfig.ClientConfig()
+	if err != nil {
+		return nil, err
+	}
+
+	t, err := rest.TransportFor(restconfig)
+	if err != nil {
+		return nil, err
+	}
+
+	// Wait for the healthz to be 200 status
+	err = wait.ForHTTPStatusOk(ctx, t, restconfig.Host+"/healthz")
+	if err != nil {
+		return nil, err
+	}
+
+	kc, err := kubernetes.NewForConfig(restconfig)
+	if err != nil {
+		return nil, err
+	}
+
+	return kc, nil
 }
